src/tool/requirements.txt
pycryptodome
pyserial
